Support for downloading unsigned dev builds.
Core packages, like `webdev/dwds`, need to run tests again dev candidate
releases that will be rolled into flutter.

This change allows users of the CI to specify whether they want to download
signed or unsigned builds. The former is what we properly call a release,
the latter is a build that is only available under the `raw` folder in our
cloud storage.

There are two other options here:
* default all dev builds to use unsigned builds. I'm not sure if this
  will break on non-linux systems, so we could do it only for linux for
  example.
* define a new name `dev_unsigned` and expose it as if it existed as a
  channel.  This would be desugared into a dev/raw request, similar to
  how the `main` name is desugared here.

Fixes #29